+++ src/components/BaseNavi.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState } from "react";
 import MenuBox from "./MenuBox";
 
@@ -35,7 +36,7 @@ export default function BaseNavi() {
           </nav>
         </div>
       </header>
-      {_switch && <MenuBox />}
+      {_switch && <MenuBox onMenuBtnSwitch={menuBtnSwitch} />}
     </>
   );
 }
