+++ src/components/MenuBox.tsx
@@ -1,14 +1,19 @@
 import React, { useRef, useState, useEffect } from "react";
 import Link from "next/link";
+import { useRouter } from "next/navigation";
 // import CircleCanvas from "./module/CircleCanvas";
 import InteractivePoints from "./module/InteractivePoints";
 
-import { useRouter } from "next/router";
 import { motion } from "framer-motion";
 
-export default function MenuBox() {
+type Props = {
+  onMenuBtnSwitch: () => void
+}
+
+export default function MenuBox(props: Props) {
+  const router = useRouter()
   const el = useRef(null);
-  const router = useRouter();
+  const onMenuBtnSwitch = props.onMenuBtnSwitch ?? (() => {})
 
   const variants = {
     open: {
@@ -55,7 +60,7 @@ export default function MenuBox() {
             className="nenu-nav__item"
           >
             <Link href="/" as="/" legacyBehavior>
-              <a className="link">Home</a>
+              <a className="link" onClick={() => onMenuBtnSwitch()}>Home</a>
             </Link>
           </motion.li>
           <motion.li
@@ -65,7 +70,7 @@ export default function MenuBox() {
             className="nenu-nav__item"
           >
             <Link href="/about" as="/about" legacyBehavior>
-              <a className="link">About</a>
+              <a className="link" onClick={() => onMenuBtnSwitch()}>About</a>
             </Link>
           </motion.li>
           <motion.li
@@ -75,7 +80,7 @@ export default function MenuBox() {
             className="nenu-nav__item"
           >
             <Link href="/memo" as="/memo" legacyBehavior>
-              <a className="link">Memo</a>
+              <a className="link" onClick={() => onMenuBtnSwitch()}>Memo</a>
             </Link>
           </motion.li>
         </motion.ul>
