--- src/pages/_app.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-import Router from "next/router";
-import "../styles/style.sass";
-
-const routeChange = () => {
-  const tempFix = () => {
-    const allStyleElems = document.querySelectorAll('style[media="x"]');
-    allStyleElems.forEach((elem) => {
-      elem.removeAttribute("media");
-    });
-  };
-  tempFix();
-};
-
-Router.events.on("routeChangeComplete", routeChange);
-Router.events.on("routeChangeStart", routeChange);
-function BaseApp({ Component, pageProps }) {
-  return (
-    <>
-      <Component {...pageProps} />
-    </>
-  );
-}
-
-export default BaseApp;
