--- src/pages/api/users.tsx
@@ -1,7 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-
-export default (req: NextApiRequest, res: NextApiResponse) => {
-  res.setHeader("content-type", "application/json");
-  res.statusCode = 200;
-  res.json({ name: "John Doe" });
-};
